+++ changelog.md
@@ -1,3 +1,4 @@
+- [2020-09-27] Remove shape check when supplying data to create plot object.
 - [2020-09-27] Fix `set_elements_manual`
 - [2020-09-17] Better guess_dimensions warnings, only warn if dimension is shared in a 
   filegroup containing the variable